# docs/swagger.yaml
basePath: /
definitions:
  api.CreateItemResponse:
    properties:
      resource:
        $ref: '#/definitions/queries.Item'
    type: object
  api.CreateSessionResponse:
    properties:
      jwt:
        type: string
      user_id:
        type: integer
    type: object
  api.GetMeResponse:
    properties:
      resource:
        $ref: '#/definitions/queries.User'
    type: object
  queries.Item:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      happened_at:
        type: string
      id:
        type: integer
      kind:
        $ref: '#/definitions/queries.Kind'
      tag_ids:
        items:
          type: integer
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  queries.Kind:
    enum:
    - expenses
    - in_come
    type: string
    x-enum-varnames:
    - KindExpenses
    - KindInCome
  queries.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: gsemir0418@gmail.com
    name: GSemir
    url: http://gsemir0418.github.com/
  description: 记账应用接口文档
  title: 记账
paths:
  /api/v1/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: 金额（单位：分）
        example: 100
        in: body
        name: amount
        required: true
        schema:
          type: integer
      - description: 类型
        example: expenses
        in: body
        name: kind
        required: true
        schema:
          type: string
      - description: 发生时间
        example: "2023-09-26T00:00:00Z"
        in: body
        name: happened_at
        required: true
        schema:
          type: string
      - description: 标签ID列表
        in: body
        name: tag_ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateItemResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - Bearer: []
      summary: 创建账目
  /api/v1/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetMeResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: 获取当前登录用户
  /api/v1/session:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮件地址
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: 验证码
        in: body
        name: code
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateSessionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: 登录
  /api/v1/validation_codes:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮件地址
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: 请求发送验证码
  /ping:
    get:
      consumes:
      - application/json
      description: 如果返回 200，说明服务正常运行
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: 服务健康度
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"